Massive refactor
Everything needed moving around, it was horrible.
Key points:

    - lib/utils has been split out into utils for different modules

    - actions are now defined using a funky action decorator that I wrote
      to try and tidy up how long the help texts are and general code
      copying.
