+++ ckanext/versioned_datastore/lib/downloads/__init__.py
@@ -1,0 +1,9 @@
+# this is a namespace package
+try:
+    import pkg_resources
+
+    pkg_resources.declare_namespace(__name__)
+except ImportError:
+    import pkgutil
+
+    __path__ = pkgutil.extend_path(__path__, __name__)
