+++ ckanext/versioned_datastore/tests/bin/test.ini
@@ -4,7 +4,7 @@ host = 0.0.0.0
 port = 5000
 
 [app:main]
-use = config:/ckan/test-core.ini
+use = config:/usr/lib/ckan/default/src/ckan/test-core.ini
 
 ckanext.versioned_datastore.elasticsearch_hosts = 127.0.0.1
 ckanext.versioned_datastore.elasticsearch_port = 9200
