+++ ckanext/versioned_datastore/tests/test_helpers.py
@@ -1,13 +1,13 @@
 from traceback import format_exception_only
 
-from ckanext.versioned_datastore.helpers import is_duplicate_ingestion, get_human_duration, \
-    get_stat_icon, get_stat_activity_class, get_stat_title
-from ckanext.versioned_datastore.lib.ingestion.exceptions import DuplicateDataSource, \
-    UnsupportedDataSource
-from ckanext.versioned_datastore.lib.stats import INGEST, INDEX, PREP, ALL_TYPES
 from ckantest.models import TestBase
 from mock import MagicMock, patch
 from nose.tools import assert_equal, assert_true, assert_false, assert_not_equal
+
+from ..helpers import is_duplicate_ingestion, get_human_duration, get_stat_icon, \
+    get_stat_activity_class, get_stat_title
+from ..lib.importing.ingestion.exceptions import DuplicateDataSource, UnsupportedDataSource
+from ..lib.importing.stats import INGEST, INDEX, PREP, ALL_TYPES
 
 
 class TestHelpers(TestBase):
